move layouts components to components
